Rename from springtxn to spring since boot support is included as well (and Environment configuration support to come)
