+++ pom.xml
@@ -9,11 +9,11 @@
   </parent>
 
   <groupId>io.ebean</groupId>
-  <artifactId>ebean-springtxn</artifactId>
-  <name>ebean-springtxn</name>
+  <artifactId>ebean-spring</artifactId>
+  <name>ebean-spring</name>
   <version>10.1.1-SNAPSHOT</version>
   <packaging>jar</packaging>
-  <description>Support for Spring transaction use with Ebean</description>
+  <description>Ebean support for Spring and Spring Boot</description>
 
   <properties>
     <spring.framework.version>4.3.4.RELEASE</spring.framework.version>
